% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{big_example_zeros}
\alias{big_example_zeros}
\title{Resulting reads that get set to zero when running `remove_false_pos(big_example)`, for testing.}
\format{A tibble with 125 rows and 23 columns:
\describe{
  \item{Sample}{Either 'mock', or 1, 2, 3, ..., 124 labelling the samples}
  \item{a1, a2, a3, a4}{Reads of control species a1, a2, a3 and a4}
  \item{`1`, `2` `3`, ..., 18}{Reads of non-control species 1, 2, 3, ..., 18}
}}
\source{
From setting false positives to zero in `big_example`.
}
\usage{
big_example_zeros
}
\description{
Expected resulting data set containing 1 mock sample, 124 non-mock samples, with
 reads for 4 control species and 18 non-control species. Gives the
 numbers in Table ????? of the manuscript, with non-zero reads indicating the
 reads that get to zero to remove false positives.
}
\keyword{datasets}
