% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_alpha.R
\name{change_alpha}
\alias{change_alpha}
\title{Loops over values of alpha -- **needs tidying up**}
\usage{
change_alpha(data, alpha_vec = seq(0, 1, 0.1), aT = 4)
}
\arguments{
\item{data}{As for remove_false_pos()}

\item{alpha_vec}{Vector of alpha (tolerance) values over which to run remove_false_pos().}

\item{aT}{Total number of control species.}
}
\value{
List consisting of:
   ***:  list of dataframes, one for each result of the corresponding
   value of alpha;
   num_zeroed: vector corresponding to the number of reads set to zero for each value of alphavec;
   num_mock_sp_left: ***
   num_samp_with_mock: ***
}
\description{
Removes false positives for various values of the tolerance parameter alpha
}
\author{
Andrew Edwards
}
