% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{small_example}
\alias{small_example}
\title{Small example data set.}
\format{A tibble with 4 rows and 8 columns:
\describe{
  \item{Sample}{Either 'mock', or 1, 2, 3, labelling the samples}
  \item{a1, a2, a3, a4}{Reads of control species a1, a2, a3 and a4}
  \item{`1`, `2` `3`}{Reads of non-control species 1, 2 and 3}
}}
\source{
Simplified and anonymised example from a real data set (big_example).
}
\usage{
small_example
}
\description{
An example data set containing one mock sample, three non-mock samples, with
 reads for four control species and three non-control species. Gives the
 numbers in Table 1 of the manuscript.
}
\keyword{datasets}
