% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_set_to_zero.R
\name{which_set_to_zero}
\alias{which_set_to_zero}
\title{Show which reads were set to zero}
\usage{
which_set_to_zero(original, result)
}
\arguments{
\item{original}{Original tibble data frame of numbers of reads.}

\item{result}{Resulting tibble from doing remove_false_pos(original).}
}
\value{
A tibble showing which of the original positive reads were set to zero.
}
\description{
Subtracts the result of remove_false_pos(original) [already done]
from original, which has to be done through this function which takes care
of the first column containing characters (the word 'mock').
}
\author{
Andrew Edwards
}
